Change logo size and relative code, change nav text size
As requested by the client, I changed the logo with a smaller
version, thus eliminating the need for smaller versions to load
on demand with srcset/sizes.

The nav text size was bigger too, so I set it to a fixed size
of 14px (smaller than the base size for the document), and 16px
for chrome-like browsers, for that browser family renders smallcaps
in a smaller font by default.